+++ commercetools.Sdk/commercetools.Sdk.Serialization/Mappers/AttributeMappers/DateAttributeMapper.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using Newtonsoft.Json.Linq;
 using Attribute = commercetools.Sdk.Domain.Products.Attributes.Attribute;
 
 namespace commercetools.Sdk.Serialization
